+++ ChangeLog.md
@@ -4,6 +4,12 @@ Note that version numbers do not necessarily reflect the amount of changes betwe
 
 Not all changes are documented here. To examine the full set of changes between versions, you can use git to browse the changes between the tags.
 
+dev version
+-----------
+
+* Add F6/F7 quick save/load support for DOSBox-X
+* Change UI buttons for quick save/load in DOSBox-X mode
+
 8.3.15 - 29.04.2015
 -------------------